Fixed a major bug with config reloading and removed an obsolete submodule
Previous implementation had a major bug where it was calling the
`HandleConfigChanged()` function directly from the event, which caused
it to be executed in another thread.
This posed an issue due to race conditions and some code not being
executed, due to certain unity functions not being available in other
threads.
This caused some of the weirdest behaviour I've ever seen, such as
functions just stopping execution midway through with no errors(at least
none displayed in the editor, they can be caught with try-catch),
unexplainable jumping when using IDE debugging etc.
New implementation simply uses a bool that is set to true if there was a
change, no matter the change type, which causes `Update()` to proceed
with the execution and reload the config.
Thanks to tupinikum and .dlich from the Official Unity Discord server
for helping me debug this one, successfully nailed by tupi.